creation of event creation page
