+++ .idea/misc.xml
@@ -3,6 +3,8 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
+        <entry key="..\:/Users/Reile/Documents/CS184/kiwi/app/src/main/res/layout/fragment_event_creation.xml" value="0.30052083333333335" />
+        <entry key="..\:/Users/Reile/Documents/CS184/kiwi/app/src/main/res/menu/bottom_nav_menu.xml" value="0.30052083333333335" />
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.19610507246376813" />
       </map>
     </option>
