+++ app/src/main/java/edu/ucsb/cs/cs184/group2/kiwi/ui/eventCreation/EventCreationFragment.kt
@@ -1,0 +1,37 @@
+package edu.ucsb.cs.cs184.group2.kiwi.ui.eventCreation
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.TextView
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.ViewModelProvider
+import edu.ucsb.cs.cs184.group2.kiwi.databinding.FragmentEventCreationBinding
+
+class EventCreationFragment : Fragment() {
+
+    private var _binding: FragmentEventCreationBinding? = null
+
+    // This property is only valid between onCreateView and onDestroyView.
+    private val binding get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        val eventCreationViewModel =
+            ViewModelProvider(this).get(EventCreationViewModel::class.java)
+
+        _binding = FragmentEventCreationBinding.inflate(inflater, container, false)
+        val root: View = binding.root
+
+        return root
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
