+++ app/src/main/java/edu/ucsb/cs/cs184/group2/kiwi/ui/eventCreation/EventCreationViewModel.kt
@@ -1,10 +1,10 @@
-package edu.ucsb.cs.cs184.group2.kiwi.ui.notifications
+package edu.ucsb.cs.cs184.group2.kiwi.ui.eventCreation
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class NotificationsViewModel : ViewModel() {
+class EventCreationViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
         value = "This is notifications Fragment"
