+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -18,8 +18,8 @@
         tools:layout="@layout/fragment_dashboard" />
 
     <fragment
-        android:id="@+id/navigation_notifications"
-        android:name="edu.ucsb.cs.cs184.group2.kiwi.ui.notifications.NotificationsFragment"
-        android:label="@string/title_notifications"
-        tools:layout="@layout/fragment_notifications" />
+        android:id="@+id/navigation_event_creation"
+        android:name="edu.ucsb.cs.cs184.group2.kiwi.ui.eventCreation.EventCreationFragment"
+        android:label="@string/title_event_creation"
+        tools:layout="@layout/fragment_event_creation" />
 </navigation>
