+++ app/src/main/res/values/strings.xml
@@ -2,5 +2,5 @@
     <string name="app_name">Kiwi</string>
     <string name="title_home">Home</string>
     <string name="title_dashboard">Dashboard</string>
-    <string name="title_notifications">Notifications!!!</string>
+    <string name="title_event_creation">Create Event</string>
 </resources>
